Move fetchWithRetry to its own module under src/shared/
fetchWithRetry doesn't depend on Express at all, so it belongs with the
other cross-cutting helpers rather than web/routes/shared.ts, which is
already a large file. Giving it its own small co-located test file also
keeps it out of the total-complexity budget of either crowded file.
