+++ src/web/routes/shared.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi, afterEach } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
 import multer from 'multer';
 import type { Response } from 'express';
 import type { SessionUser } from '../../types/express';
@@ -39,7 +39,6 @@ import {
   parseRewardIdParam,
   handleReservedOrConflictCommandError,
   createMulterErrorRedirectHandler,
-  fetchWithRetry,
 } from './shared';
 
 describe('parseCheckboxField', () => {
@@ -590,70 +589,5 @@ describe('createMulterErrorRedirectHandler', () => {
     expect(handler(err, res)).toBe(true);
     expect(error).toHaveBeenCalledWith('Overlay upload middleware error:', err);
     expect(redirect).toHaveBeenCalledWith('/overlay/settings?error=upload_failed');
-  });
-});
-
-describe('fetchWithRetry', () => {
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  function mockLog() {
-    const warn = vi.fn();
-    return { log: { warn } as unknown as import('winston').Logger, warn };
-  }
-
-  function mockResponses(responses: { ok: boolean; status?: number }[]) {
-    let callIndex = 0;
-    return vi.spyOn(globalThis, 'fetch').mockImplementation(() => {
-      const r = responses[callIndex++];
-      return Promise.resolve({ ok: r.ok, status: r.status ?? 200 } as globalThis.Response);
-    });
-  }
-
-  it('returns the response immediately when it is ok, without retrying', async () => {
-    const fetchSpy = mockResponses([{ ok: true, status: 200 }]);
-    const { log, warn } = mockLog();
-
-    const res = await fetchWithRetry('https://example.com', {}, log);
-
-    expect(res.ok).toBe(true);
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-    expect(warn).not.toHaveBeenCalled();
-  });
-
-  it('does not retry a non-transient error status', async () => {
-    const fetchSpy = mockResponses([{ ok: false, status: 400 }]);
-    const { log, warn } = mockLog();
-
-    const res = await fetchWithRetry('https://example.com', {}, log);
-
-    expect(res.ok).toBe(false);
-    expect(res.status).toBe(400);
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-    expect(warn).not.toHaveBeenCalled();
-  });
-
-  it('retries a transient 503 and returns the successful response', async () => {
-    const fetchSpy = mockResponses([{ ok: false, status: 503 }, { ok: true, status: 200 }]);
-    const { log, warn } = mockLog();
-
-    const res = await fetchWithRetry('https://example.com', {}, log, 2, 1);
-
-    expect(res.ok).toBe(true);
-    expect(fetchSpy).toHaveBeenCalledTimes(2);
-    expect(warn).toHaveBeenCalledTimes(1);
-  });
-
-  it('gives up and returns the last response after exhausting retries', async () => {
-    const fetchSpy = mockResponses([{ ok: false, status: 503 }, { ok: false, status: 503 }, { ok: false, status: 503 }]);
-    const { log, warn } = mockLog();
-
-    const res = await fetchWithRetry('https://example.com', {}, log, 2, 1);
-
-    expect(res.ok).toBe(false);
-    expect(res.status).toBe(503);
-    expect(fetchSpy).toHaveBeenCalledTimes(3);
-    expect(warn).toHaveBeenCalledTimes(2);
   });
 });
